Refactor to improve code readability
