+++ app/src/components/Navbar.jsx
@@ -51,22 +51,6 @@ export default function Navbar() {
     </Disclosure.Button>
   ));
 
-  const renderUserNavigation = userNavigation.map((item) => (
-    <Menu.Item key={item.name}>
-      {({ active }) => (
-        <a
-          href={item.href}
-          className={classNames(
-            active ? "bg-gray-100" : "",
-            "block px-4 py-2 text-sm text-gray-700",
-          )}
-        >
-          {item.name}
-        </a>
-      )}
-    </Menu.Item>
-  ));
-
   return (
     <Disclosure as='nav' className='bg-cyan-800'>
       {({ open }) => (
@@ -122,7 +106,7 @@ export default function Navbar() {
   );
 }
 
-function classNames(...classes) {
+export function classNames(...classes) {
   return classes.filter(Boolean).join(" ");
 }
