+++ app/src/components/user/Login.jsx
@@ -1,40 +1,11 @@
-import React, { Fragment, useState, useEffect, useContext } from "react";
+import React, { Fragment, useEffect, useContext } from "react";
 import { Menu, Transition } from "@headlessui/react";
 import { users, UserContext } from "./UserContext";
-
+import UserNavigation from "./UserNavigation";
 
 export default function Login() {
-  const [userNavigation, setUserNavigation] = useState(
-    Object.values(users)
-  )
 
   const [user, setUser] = useContext(UserContext);
-
-  const renderUserNavigation = userNavigation.map((item) => (
-    <Menu.Item 
-      onClick={() => {
-        const target = item.name;
-        if (target === "Logout") {
-          setUser("Login")
-          sessionStorage.removeItem("user");
-        } else {
-          setUser(target);
-          sessionStorage.setItem("user", target);
-        }
-      }}
-      key={item.key}>
-      {({ active }) => (
-        <p
-          className={classNames(
-            active ? "bg-gray-100" : "",
-            "block px-4 py-2 text-sm text-gray-700",
-          )}
-        >
-          {item.name}
-        </p>
-      )}
-    </Menu.Item>
-  ));
 
   useEffect(() => {
     if (sessionStorage.getItem("user") != null && sessionStorage.getItem("user").toLowerCase() in users) {
@@ -62,13 +33,9 @@ export default function Login() {
         leaveTo='transform opacity-0 scale-95'
       >
         <Menu.Items className='absolute right-0 z-10 mt-2 w-48 origin-top-right rounded-md bg-white py-1 shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none'>
-          {renderUserNavigation}
+          <UserNavigation/>
         </Menu.Items>
       </Transition>
     </Menu>
   )
-}
-
-function classNames(...classes) {
-  return classes.filter(Boolean).join(" ");
 }
